fixed bug in the cluster code.
