+++ experiments/cluster_scripts/numeric_experiments_sbatch_runner.py
@@ -13,11 +13,11 @@ from experiments.cluster_scripts.common import (
     create_all_experiments_folders,
     EXPERIMENTS_CONFIG_STR,
     submit_job_and_validate_execution,
+    create_execution_arguments,
 )
-from experiments.cluster_scripts.multi_agent_experiments_sbatch_runner import create_execution_arguments
-from utilities import LearningAlgorithmType
 
 signal.signal(signal.SIGINT, sigint_handler)
+learning_algorithms_map = {3: "nsam", 15: "naive_nsam"}
 
 
 def execute_statistics_collection_job(code_directory, configuration, environment_variables, experiment, job_ids, internal_iterations):
@@ -77,7 +77,7 @@ def main():
                     experiment_sids.append(sid)
                     formatted_date_time = datetime.now().strftime("%A, %B %d, %Y %I:%M %p")
                     print(
-                        f"{formatted_date_time} - submitted job with sid {sid} for algorithm {LearningAlgorithmType(version_index).name} fold {fold} and iteration {internal_iteration}."
+                        f"{formatted_date_time} - submitted job with sid {sid} for algorithm {learning_algorithms_map[version_index]} fold {fold} and iteration {internal_iteration}."
                     )
                     pathlib.Path("temp.sbatch").unlink()
                     progress_bar(version_index, len(experiment["compared_versions"]))
@@ -97,7 +97,7 @@ def main():
                 )
                 formatted_date_time = datetime.now().strftime("%A, %B %d, %Y %I:%M %p")
                 print(
-                    f"{formatted_date_time} - submitted job to run experiment for triplets with sid {sid} for algorithm {LearningAlgorithmType(version_index).name} and fold {fold}."
+                    f"{formatted_date_time} - submitted job to run experiment for triplets with sid {sid} for algorithm {learning_algorithms_map[version_index]} and fold {fold}."
                 )
 
             time.sleep(5)
